Ordenamiento del código y estructura blueprint, modularización de templates and styles
