+++ app/main/templates/cards/course_card.html
