+++ app/main/templates/index.html
@@ -1,4 +1,4 @@
-{% extends 'partials/base.html' %} {% block title %} Inicio - CorEngineer {% endblock %}
+{% extends 'base.html' %} {% block title %} Inicio - CorEngineer {% endblock %}
 {% block contentBlock %}
 <div class="index">
   <div class="index__background"></div>
