+++ app/main/templates/partials/footer.html
@@ -1,0 +1,5 @@
+<footer class="footer">
+    <div class="footer__container">
+        <p>&copy; 2024 CORENGINEER. Todos los derechos reservados.</p>
+    </div>
+</footer>
