+++ app/main/templates/partials/links.html
@@ -19,12 +19,3 @@
 <link rel="stylesheet" href="{{ url_for('static', filename='css/cards/service_card.css') }}">
 
 <link rel="stylesheet" href="{{ url_for('static', filename='css/cards/profile_card.css') }}">
-
-
-
-
-
-
-
-
-
