+++ app/main/templates/partials/metadata.html
@@ -1,0 +1,35 @@
+<!-- metadata.html -->
+<meta charset="UTF-8">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta name="description" content="Descripción breve de tu página web">
+<meta name="keywords" content="palabras, clave, separadas, por, comas">
+<meta name="author" content="Tu nombre o el nombre de la compañía">
+<meta name="robots" content="index, follow">
+
+<!-- Open Graph -->
+<meta property="og:title" content="Título de la página para redes sociales">
+<meta property="og:description" content="Descripción de la página para redes sociales">
+<meta property="og:image" content="URL de la imagen para redes sociales">
+<meta property="og:url" content="URL de la página">
+<meta property="og:type" content="website">
+
+<!-- Twitter Cards -->
+<meta name="twitter:card" content="summary_large_image">
+<meta name="twitter:title" content="Título para Twitter">
+<meta name="twitter:description" content="Descripción para Twitter">
+<meta name="twitter:image" content="URL de la imagen para Twitter">
+<meta name="twitter:site" content="@TuCuentaDeTwitter">
+
+<!-- SEO y Compatibilidad -->
+<meta http-equiv="X-UA-Compatible" content="IE=edge">
+
+<!-- Seguridad y Privacidad -->
+ <!--
+<meta http-equiv="Content-Security-Policy" content="default-src 'self'; script-src 'self' https://example.com;">
+<meta name="referrer" content="no-referrer">
+-->
+
+<!-- Colores y Apariencia -->
+<meta name="theme-color" content="#FFFFFF">
+<meta name="msapplication-TileColor" content="#FFFFFF">
+<meta name="msapplication-TileImage" content="/path/to/tile-image.png">
