+++ app/main/templates/services.html
@@ -1,4 +1,4 @@
-{% extends 'partials/base.html' %} {% block title %}Services CorEngineer{% endblock %} 
+{% extends 'base.html' %} {% block title %}Services CorEngineer{% endblock %} 
 
 
 {%block contentBlock %}
