--- app/templates/partials/base.html
@@ -1,73 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>{% block title%} {% endblock %}</title>
-        {% include 'partials/links.html' %}
-    </head>
-    <body>
-        <header class="header">
-                <!--Logo header-->
-                <a href="{{url_for('main.home')}}" class="header__link-logo">
-                    <div class="header__logo-container">
-                        <img src="{{ url_for('static', filename='img/logo/logo_inskape.svg') }}" alt="logo"
-                            class="header__logo-img">
-                    </div>
-                    <h3 class="header__logo-text">CorEngineer</h3>
-                </a>
-
-                <input type="checkbox" class="header__nav-checkbox-menu" id="menu-bars" role="button">
-                <label for="menu-bars" class="header__nav-menu-button" role="button" alt="menu">
-                    <i class="header__nav-button-ico fa-solid fa-bars"></i>
-                </label>
-
-                <nav class="header__nav">
-                    <ul class="header__nav-list">
-                        <li class="header__nav-item">
-                            <a href="{{url_for('main.home')}}" class="header__nav-link">
-                                <i class="header__nav-item-ico fa-solid fa-house"></i><h5 class="header__nav-item-text">home</h5 >
-                            </a>
-                        </li>
-                        <li class="header__nav-item">
-                            <a href="{{url_for('main.services')}}" class="header__nav-link">
-                                <i class="header__nav-item-ico fas fa-cogs"></i><h5 class="header__nav-item-text">services</h5 >
-                            </a>
-                        </li>
-                        <li class="header__nav-item">
-                            <a href="{{url_for('main.portfolio')}}" class="header__nav-link">
-                                <i class="header__nav-item-ico fa-sharp fa-solid fa-briefcase"></i><h5 class="header__nav-item-text">portfolio </h5 >
-                            </a>
-                        </li>
-                        <li class="header__nav-item">
-                            <a href="{{url_for('main.team')}}" class="header__nav-link">
-                                <i class="header__nav-item-ico fa-solid fa-user-group"></i><h5 class="header__nav-item-text">team</h5 >
-                            </a>
-                        </li>
-                        <li class="header__nav-item">
-                            <a href="{{url_for('main.contact')}}" class="header__nav-link">
-                                <i class="header__nav-item-ico fa-solid fa-address-book"></i><h5 class="header__nav-item-text">contact</h5 >
-                            </a>
-                        </li>
-                        <li class="header__nav-item">
-                            <a href="{{url_for('main.about')}}" class="header__nav-link">
-                                <i class="header__nav-item-ico fa-solid fa-circle-info"></i><h5 class="header__nav-item-text">about </h5 >
-                            </a>
-                        </li>
-                    </ul>
-                </nav>   
-        </header>
-        
-        <main class="main">
-            {% block contentBlock %}{% endblock %}
-        </main>
-        
-        <script src="{{url_for('static', filename='js/main.js')}}" type="module"></script>
-    </body>
-
-    <footer class="footer">
-        <div class="footer__container">
-            <p>&copy; 2024 CORENGINEER. Todos los derechos reservados.</p>
-        </div>
-    </footer>
-</html>
