+++ static/css/cards/course_card.css
